Thread suspension: Allow natp->suspended_registers' type to vary by platform
This allows us to clean up the Windows implementation so that we don't
have to re-call GetThreadContext() to avoid resetting extended state
that is not covered by NaClSignalContext (e.g. SSE registers and FP
state).

This is in preparation for implementing
NaClAppThreadSetSuspendedRegisters() on Mac OS X, where we will want
to store additional state besides NaClSignalContext.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2879
TEST=run_thread_suspension_test

Review URL: https://chromiumcodereview.appspot.com/10824362

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9508 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2