Thread suspension: Allow modifying registers on Mac OS X
This is tricky on x86-32 because Mac's thread_set_state() call resets
%cs back to the trusted-code value.  We reuse the NaClSwitchAllRegs()
code to work around this.

Move the definition of "struct NaClSwitchAllRegsState" into a header
so that osx/thread_suspension.c can allocate an instance of it.
Similarly, split NaClSwitchAllRegsSetup() out from NaClSwitchAllRegs()
so that the Mac code can call it.

Change the Mac code to restore control to untrusted code via a subset
of the NaClSwitchAllRegsAsm routine.

Add a missing 'const' to NaClSwitchAllRegs().

Enable the register modification test to run on Mac.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2879
TEST=run_thread_suspension_test

Review URL: https://chromiumcodereview.appspot.com/10825433

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9520 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2