Add test for the faulted thread queue (which is used by the debug stub)
Before this change, the faulted thread queue is only tested via the
debug stub's use of it.  However, it's useful to have a direct test.
This is in preparation for implementing the faulted thread queue on
Mac OS X, because implementing this is not yet sufficient to turn on
the debug stub tests for Mac.

The test checks that a faulted thread is reported properly and that
the thread can be continued.  On x86, it also tests single-stepping.

We do some cleanups to make it easier to test this without duplicating
too much code:

 * Split out NaClFaultedThreadQueueEnable() from nacl_debug.cc so that
   our test can use it

 * Split out NaClDebugExceptionHandlerStandaloneHandleArgs() from
   sel_main.c so that our test can use it.

 * Define NACL_ABI_SIG* as unconditionally #defined
   platform-independent signal numbers that the debug stub reports to
   its clients.

 * Make win/debug_exception_handler.h unconditionally #includable (as
   osx/mach_exception_handler.h is), and make
   NaClDebugExceptionHandlerEnsureAttached() unconditionally callable,
   to reduce the number of #ifs.

 * Extend RegsAssertEqual() to check the x86 trap flag.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2879
TEST=run_faulted_thread_queue_test

Review URL: https://chromiumcodereview.appspot.com/10876021

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9547 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2