Fix breakage of GDB tests on the ASan bots, introduced by r9556
r9556 changed the debug stub to use SIGSEGV for the initial breakpoint
rather than SIGTRAP.  However, ASan normally intercepts SIGSEGV (but
not SIGTRAP), which broke the GDB tests -- and we don't have a trybot
which covers ASan to catch this.

We can fix this by telling ASan not to intercept SIGSEGV with
"ASAN_OPTIONS=handle_segv=0".  This means we can re-enable the
platform qualification test in the GDB tests, and enable the debug
stub tests too.

We could enable more tests under ASan this way, but I want to keep
this change small since it's fixing build breakage.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2879
BUG=http://code.google.com/p/nativeclient/issues/detail?id=2717
TEST="large_tests --asan" (covers debug stub and GDB tests; no trybots run this)

Review URL: https://chromiumcodereview.appspot.com/10876068

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9563 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2