Make NaClLogModuleInit() idempotent
This is to allow calling NaClDescMakeCustomDesc() before
NaClChromeMainStart(), which the new Chrome IPC-based PPAPI proxy
needs to be able to do.  Currently the new proxy crashes on startup on
Windows and Mac because NaClLog's mutex is not initialised when
NaClDescMakeCustomDesc() causes a call to NaClLog().

We change sel_main_chrome.c so that NaClChromeMainArgsCreate()
implicitly initialises NaClLog.  Doing it implicitly in this way is a
little hacky, but it saves us from having to expose another
initialisation function to Chromium and from having to change Chromium
to use it.  We could add such a function later if it turns out to be
necessary or cleaner.

Extend sel_main_chrome_test.cc to test usage of
NaClDescMakeCustomDesc().

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1696
TEST=run_sel_main_chrome_test

Review URL: https://chromiumcodereview.appspot.com/10894028

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9599 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2