Scons: Remove the obsolete buildplatform and targetplatform options
The last use of these options (in the PNaCl toolchain build) was
removed recently in r9597.  These options are not tested by the
trybots and are prone to get broken.  I'm removing them so that no-one
starts depending on them again.  This will also enable further
cleanups.

The original purpose of these options was for cross-compiling
validators, e.g. to allow the ARM validator to be built to run on an
x86 system.  But these days we handle that by building arm-ncval-core
as part of the normal x86-32 build without requiring any special Scons
options.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2703
TEST=trybots

Review URL: https://chromiumcodereview.appspot.com/10914021

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9610 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2