Debug stub: Fix potential race condition in startup
When GDB connects to the debug stub, it expects the process to already
be suspended.  GDB issues a "?" command to get the thread's status,
and we want this to report that the process stopped at the initial
breakpoint.

However, there is a race condition that means that might not always
happen.  If the initial thread is slow to start up, the debug stub
might receive and respond to GDB's message before the initial thread
reaches the initial breakpoint.

This can cause tests in debug_stub_test.py fail.  This can be observed
by adding "sleep(1)" into NaClThreadLauncher() in nacl_app_thread.c,
though I've not seen the failures occur without that change.

We fix this by waiting for the breakpoint to occur before handling
messages from GDB.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2983
TEST=run_debug_stub_test with sleep call added to reproduce problem;
  no test added because it's difficult to test a race condition

Review URL: https://chromiumcodereview.appspot.com/10916062

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9629 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2