Do not require IRT writable data segment to be page-aligned
This matches the change made in r8781 for the main nexe loading.
It applies the same loosening of requirements to the loading of
the IRT ("blob library").

ELF specifies that p_vaddr need not be aligned, just that
(p_vaddr & p_align) == (p_offset & p_align).  The normal layout
places the writable data immediately after the rodata in the file,
and at edata+PAGE_SIZE in the memory image.  This avoids bloating
the file size by up to PAGE_SIZE-1.

The next-generation linker for NaCl uses this layout too.
This changes sel_ldr to accept that.  It has the same semantics
as the usual ELF mmap-based loading: the partial page before the
p_vaddr address is loaded from the file too.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2814
TEST= trybots and new run_unaligned_data_irt_test

R=bsy@google.com

Review URL: https://codereview.chromium.org/10983053

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9861 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2