Enable the "-Wundef" warning for non-PNaCl untrusted code
-Wundef warns about "#if FOO" when FOO is undefined.

I've had to filter out -Wundef in one case.

This fixes some assembly code to build with -Wundef.  We are not
changing the behaviour here because EXPECTED_RODATA is never #defined
to be 0.  See service_runtime/nacl.scons: A rodata_address value of
None is mapped to EXPECTED_RODATA being undefined, while a value of 0
gets mapped to a non-zero value of EXPECTED_RODATA.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2787
TEST=build

Review URL: https://codereview.chromium.org/11000035

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9879 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2