Add a test for __attribute__((constructor))/__attribute__((destructor))
This is preparation for changing how PNaCl implements
constructors/destructors.

The existing test that uses __attribute__((section(".init_array")))
will not work after we do the PNaCl change -- this test is too
low-level and does not test the construct that programs actually use
-- so we want a test to replace it.

In implementing this, I discovered that __attribute__((destructor)) is
broken in x86-64 nacl-glibc.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3018
TEST=run_initfini_attributes_test

Review URL: https://codereview.chromium.org/11027013

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9905 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2