Fix movslq handling.
The movslq instruction is not marked as nacl-amd64-zero-extends instruction
in R-DFA (while old validator accepts it as such).  Additionally error message
is wrong: it says base is wrong when index is wrong and issues cryptic
"DFA error in validator" message when DFA does not recognize instruction - but
this message implies that DFA is somehow wrong, which is, of course, not true.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3060
TEST=trybots
Review URL: https://codereview.chromium.org/11044015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9917 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2