service_runtime: Simplify by removing non-__PIC__ code paths
All of the "#ifdef __PIC__" cases work regardless of whether we are
building a position-independent executable or library.  We might as
well simplify the code by removing the non-__PIC__ code paths.

The non-__PIC__ code paths are currently dead code: they are not
tested on the NaCl or Chromium bots, which always build with -fPIC or
-fPIE.

The non-__PIC__ code paths provide a tiny micro-optimisation by using
shorter code sequences, but having non-position-independent code is
riskier from a security point of view, so it's not worth keeping code
around for the (probably unmeasurable) speedup.

Furthermore, some of this non-PIC code fails to build with -Wundef
because it uses "#if __PIC__" rather than "#ifdef __PIC__".  (We
discovered this on x86-32 because Clang fails to define __PIC__ when
given -fPIE.)

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3059
TEST=trybots

Review URL: https://codereview.chromium.org/11046026

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9926 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2