Debug stub tests: Test that calling a NULL function pointer is reported
For some of the thread suspension work, I want to change
NaClSignalContextIsUntrusted() to return false when prog_ctr points
into a trampoline sequence.  The naive change there would be for it to
check "prog_ctr >= NACL_TRAMPOLINE_END", but this would have the side
effect of stopping the debug stub from reporting jump-to-address-zero
crashes to the debugger on Linux.

In order to prevent such a regression, I'm adding this debug stub
test.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2830
TEST=run_debug_stub_test

Review URL: https://codereview.chromium.org/11029057

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9938 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2