Add a getpid() dev interface to the IRT
Before this change, libnacl contains a getpid() implementation which
does a direct NaCl syscall.  Since this libnacl is shipped in the NaCl
SDK, programs are likely to be linking against this code and calling
it, which is unfortunate because this potentially gives undefined
behaviour.  We are not guaranteeing that calling a NaCl syscall
trampoline directly will not crash the program in the future.

This change allows us to clean up libnacl so that it no longer
contains this direct syscall.

This new IRT interface is only intended to do something useful when
used outside of Chromium.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2488
TEST=run_getpid_test{,_irt}

Review URL: https://codereview.chromium.org/11036066

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9941 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2