Move gdb_tests to medium_tests; make them work when "tests_use_irt" is set
Moving gdb_tests from large_tests to medium_tests is preparation for
enabling the GDB tests on ARM, because large_tests is not always
treated the same as the other test suites: large_tests is not always
run fully for ARM on the PNaCl bots, and large_tests_irt is currently
not run on any bots.

In order to do this move, we need to make gdb_tests work when
tests_use_irt is set (which was not tested on the bots before).
run_gdb_syscall_thread_test_irt failed on x86-32 with the following
error:

  Traceback (most recent call last): ...
    File "tests/gdb/syscall_thread.py", line 15, in CheckBacktrace
      assert s2 in s1, '%s not in %s' % (functions, all_functions)
  AssertionError: ['pthread_join', 'test_syscall_thread'] not in ['??', 'main']

This was because GDB wasn't being given the IRT, so it couldn't unwind
the stack from the syscall.  We fix this by using the "nacl-irt"
command.

I've also cleaned up the test so that we don't load the IRT when
tests_use_irt is not set.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3092
TEST=gdb_tests{,_irt} --mode=dbg-host,nacl,nacl_irt_test

Review URL: https://codereview.chromium.org/11151020

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10020 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2