Rename NaClPostTestInjectionMain() to NaClSelLdrMain() for use by embedders
This is intended to be a (fairly) stable interface for embedders of
NaCl to use so that embedders can link a custom version of sel_ldr
that does some custom initialisation in main() before calling into
NaClSelLdrMain().

We move the prototype of NaClSelLdrMain() into a new header file,
sel_main.h, so that embedders can #include this without pulling in a
load of NaCl internal stuff from sel_ldr.h.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3098
TEST="run_hello_world_test run_nacl_thread_capture_test" cover this

Review URL: https://codereview.chromium.org/11288004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10105 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2