Add GYP support for translating a pexe to a nexe, and testing it
This Cl adds GYP actions in untrusted.gypi for translating a pexe to a nexe
(enabled via the build_pnacl_newlib variable) and an action in tests.gyp for
testing hello_world, as well as the requiredsupport in build_nexe.py
and test_build.py.

Because of GYP's limitations (i.e. not having a separate environment for
building the IRT), this is currently the only way to build working nexes
for ARM in gyp (the flags used in the 'native' arm nexes are required to build
the IRT but do not work for normal nexes).

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2484

Review URL: https://codereview.chromium.org/11343019

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10151 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2