Enable the -Wstrict-prototypes warning for trusted code (Scons and Gyp builds)
This warning/error requires defining functions as "foo(void)" in C
rather than "foo()" because, in C (but not C++), the latter defines a
function with unspecified arguments rather than no arguments.  Using
"foo()" means that some argument mismatch mistakes will not be caught
by the compiler.

Note that this does not cover the Mac Gyp build because I don't know
how to add a C-only option in that case.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3114
TEST=build

Review URL: https://codereview.chromium.org/11377057

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10243 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2